+++ tests/runtime/test_grpc_server.py
@@ -64,7 +64,7 @@ from sample_lib.data_model import (
     SampleOutputType,
     SampleTrainingType,
 )
-from tests.conftest import random_test_id, temp_config
+from tests.conftest import random_test_id
 from tests.fixtures import Fixtures
 from tests.runtime.conftest import runtime_grpc_test_server
 import caikit
@@ -777,7 +777,7 @@ def test_canceling_model_loads_causes_exceptions(runtime_grpc_server):
         assert request_finished.is_set()
 
 
-def test_tls(sample_inference_service):
+def test_tls(sample_inference_service, open_port):
     """Boot up a server with TLS enabled and ping it on a secure channel"""
     ca_key = tls_test_tools.generate_key()[0]
     ca_cert = tls_test_tools.generate_ca_cert(ca_key)
@@ -787,6 +787,7 @@ def test_tls(sample_inference_service):
         server=KeyPair(cert=tls_cert, key=tls_key), client=KeyPair(cert="", key="")
     )
     with runtime_grpc_test_server(
+        open_port,
         inference_service=sample_inference_service,
         training_service=None,
         tls_config_override=tls_config,
@@ -794,7 +795,7 @@ def test_tls(sample_inference_service):
         _assert_connection(_make_secure_channel(server, ca_cert))
 
 
-def test_mtls(sample_inference_service):
+def test_mtls(sample_inference_service, open_port):
     """Boot up a server with mTLS enabled and ping it on a secure channel"""
     ca_key = tls_test_tools.generate_key()[0]
     ca_cert = tls_test_tools.generate_ca_cert(ca_key)
@@ -804,6 +805,7 @@ def test_mtls(sample_inference_service):
         server=KeyPair(cert=tls_cert, key=tls_key), client=KeyPair(cert=ca_cert, key="")
     )
     with runtime_grpc_test_server(
+        open_port,
         inference_service=sample_inference_service,
         training_service=None,
         tls_config_override=tls_config,
@@ -821,7 +823,7 @@ def test_mtls(sample_inference_service):
             stub.Check(health_check_request)
 
 
-def test_certs_can_be_loaded_as_files(sample_inference_service, tmp_path):
+def test_certs_can_be_loaded_as_files(sample_inference_service, tmp_path, open_port):
     """mTLS test with all tls configs loaded from files"""
     ca_key = tls_test_tools.generate_key()[0]
     ca_cert = tls_test_tools.generate_ca_cert(ca_key)
@@ -844,6 +846,7 @@ def test_certs_can_be_loaded_as_files(sample_inference_service, tmp_path):
         client=KeyPair(cert=ca_cert_path, key=""),
     )
     with runtime_grpc_test_server(
+        open_port,
         inference_service=sample_inference_service,
         training_service=None,
         tls_config_override=tls_config,
@@ -855,11 +858,12 @@ def test_certs_can_be_loaded_as_files(sample_inference_service, tmp_path):
 
 
 def test_metrics_stored_after_server_interrupt(
-    sample_task_model_id, sample_inference_service
+    sample_task_model_id, sample_inference_service, open_port
 ):
     """This tests the gRPC server's behaviour when interrupted"""
 
     with runtime_grpc_test_server(
+        open_port,
         inference_service=sample_inference_service,
         training_service=None,
     ) as server:
@@ -890,27 +894,6 @@ def test_metrics_stored_after_server_interrupt(
             assert data[0]["model_type_counters"] == {
                 "<class 'sample_lib.modules.sample_task.sample_implementation.SampleModule'>": 1
             }
-
-
-def test_out_of_range_port(sample_inference_service):
-    """Test that the server can use a port outside of the default 8888-9000
-    range
-    """
-    free_high_port = RuntimeGRPCServer._find_port(50000, 60000)
-    with temp_config(
-        {
-            "runtime": {
-                "port": free_high_port,
-                "find_available_port": False,
-            }
-        },
-        merge_strategy="merge",
-    ):
-        with runtime_grpc_test_server(
-            inference_service=sample_inference_service,
-            training_service=None,
-        ) as server:
-            _assert_connection(grpc.insecure_channel(f"localhost:{free_high_port}"))
 
 
 def test_reflection_enabled(runtime_grpc_server):
